Move PythonSetup to src/python/pants/python.
