+++ contrib/python/src/python/pants/contrib/python/checks/tasks/checkstyle/BUILD
@@ -5,7 +5,6 @@ python_library(
   dependencies=[
     '3rdparty/python:pex',
     'contrib/python/src/python/pants/contrib/python/checks/checker',
-    'src/python/pants/backend/python/subsystems',
     'src/python/pants/backend/python/targets',
     'src/python/pants/backend/python/tasks',
     'src/python/pants/backend/python:python_requirement',
