+++ contrib/python/tests/python/pants_test/contrib/python/checks/tasks/BUILD
@@ -5,7 +5,6 @@ python_tests(
   name='python_eval',
   sources=['test_python_eval.py'],
   dependencies=[
-    'src/python/pants/backend/python/subsystems',
     'src/python/pants/python',
     'contrib/python/src/python/pants/contrib/python/checks/tasks',
     'tests/python/pants_test/backend/python/tasks:python_task_test_base',
