+++ src/python/pants/backend/python/lint/isort/rules.py
@@ -15,7 +15,6 @@ from pants.backend.python.rules.pex import (
   PexRequirements,
 )
 from pants.backend.python.subsystems import python_native_code, subprocess_environment
-from pants.backend.python.subsystems.python_setup import PythonSetup
 from pants.backend.python.subsystems.subprocess_environment import SubprocessEncodingEnvironment
 from pants.engine.fs import Digest, DirectoriesToMerge, PathGlobs, Snapshot
 from pants.engine.isolated_process import (
@@ -28,6 +27,7 @@ from pants.engine.rules import UnionRule, rule, subsystem_rule
 from pants.engine.selectors import Get
 from pants.option.custom_types import GlobExpansionConjunction
 from pants.option.global_options import GlobMatchErrorBehavior
+from pants.python.python_setup import PythonSetup
 from pants.rules.core.fmt import FmtResult
 from pants.rules.core.lint import LintResult
