+++ src/python/pants/backend/python/tasks/python_execution_task_base.py
@@ -8,7 +8,6 @@ from typing import Dict, Optional
 from pex.interpreter import PythonInterpreter
 from pex.pex import PEX
 
-from pants.backend.python.subsystems.python_setup import PythonSetup
 from pants.backend.python.targets.python_distribution import PythonDistribution
 from pants.backend.python.targets.python_requirement_library import PythonRequirementLibrary
 from pants.backend.python.targets.python_target import PythonTarget
@@ -17,6 +16,7 @@ from pants.backend.python.tasks.resolve_requirements import ResolveRequirements
 from pants.backend.python.tasks.resolve_requirements_task_base import ResolveRequirementsTaskBase
 from pants.build_graph.files import Files
 from pants.invalidation.cache_manager import VersionedTargetSet
+from pants.python.python_setup import PythonSetup
 from pants.util.contextutil import temporary_file
