+++ src/python/pants/backend/python/tasks/resolve_requirements_task_base.py
@@ -10,11 +10,11 @@ from pex.pex_builder import PEXBuilder
 
 from pants.backend.python.python_requirement import PythonRequirement
 from pants.backend.python.subsystems.python_native_code import PythonNativeCode
-from pants.backend.python.subsystems.python_setup import PythonSetup
 from pants.backend.python.targets.python_requirement_library import PythonRequirementLibrary
 from pants.base.hash_utils import hash_all
 from pants.invalidation.cache_manager import VersionedTargetSet
 from pants.python.pex_build_util import PexBuilderWrapper
+from pants.python.python_setup import PythonSetup
 from pants.task.task import Task
 from pants.util.dirutil import safe_concurrent_creation
 from pants.util.memo import memoized_property
