+++ src/python/pants/python/BUILD
@@ -8,6 +8,7 @@ python_library(
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
     'src/python/pants/build_graph',
+    'src/python/pants/option',
     'src/python/pants/subsystem',
     'src/python/pants/util:collections',
     'src/python/pants/util:contextutil',
@@ -25,6 +26,7 @@ python_tests(
     'src/python/pants/build_graph',
     'src/python/pants/util:contextutil',
     'src/python/pants/testutil/subsystem',
+    'src/python/pants/testutil:pexrc_util',
     'src/python/pants/testutil:test_base',
   ],
   tags = {'partially_type_checked'},
