+++ src/python/pants/python/pex_build_util.py
@@ -4,7 +4,7 @@
 import logging
 import os
 from pathlib import Path
-from typing import Sequence, Set, Callable
+from typing import Callable, Sequence, Set
 
 from pex.fetcher import Fetcher
 from pex.pex_builder import PEXBuilder
@@ -13,12 +13,12 @@ from pex.util import DistributionHelper
 from twitter.common.collections import OrderedSet
 
 from pants.backend.python.python_requirement import PythonRequirement
-from pants.backend.python.subsystems.python_setup import PythonSetup
 from pants.base.build_environment import get_buildroot
 from pants.base.exceptions import TaskError
 from pants.build_graph.files import Files
 from pants.build_graph.target import Target
 from pants.python.python_repos import PythonRepos
+from pants.python.python_setup import PythonSetup
 from pants.subsystem.subsystem import Subsystem
 from pants.util.collections import assert_single_element
 from pants.util.contextutil import temporary_file
