--- tests/python/pants_test/backend/python/subsystems/BUILD
@@ -1,14 +1,0 @@
-# Copyright 2018 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-python_tests(
-  name='python_setup',
-  sources=['test_python_setup.py'],
-  dependencies=[
-    'src/python/pants/backend/python/subsystems',
-    'src/python/pants/util:contextutil',
-    'src/python/pants/testutil:test_base',
-    'src/python/pants/testutil:pexrc_util',
-  ],
-  tags = {"partially_type_checked"},
-)
