+++ tests/python/pants_test/backend/python/tasks/BUILD
@@ -47,10 +47,10 @@ python_tests(
   dependencies = [
     '3rdparty/python:pex',
     'src/python/pants/backend/python:interpreter_cache',
-    'src/python/pants/backend/python/subsystems',
     'src/python/pants/backend/python/targets',
     'src/python/pants/backend/python/tasks',
     'src/python/pants/build_graph',
+    'src/python/pants/python',
     'src/python/pants/source',
     'src/python/pants/util:contextutil',
     'src/python/pants/testutil:task_test_base',
@@ -222,10 +222,10 @@ python_tests(
     '3rdparty/python:pex',
     'src/python/pants/backend/python:interpreter_cache',
     'src/python/pants/backend/python:python_requirement',
-    'src/python/pants/backend/python/subsystems',
     'src/python/pants/backend/python/targets',
     'src/python/pants/backend/python/tasks',
     'src/python/pants/base:build_environment',
+    'src/python/pants/python',
     'src/python/pants/task',
     'src/python/pants/util:contextutil',
     'src/python/pants/testutil:task_test_base',
@@ -239,10 +239,10 @@ python_tests(
   dependencies = [
     '3rdparty/python:pex',
     'src/python/pants/backend/python:interpreter_cache',
-    'src/python/pants/backend/python/subsystems',
     'src/python/pants/backend/python/targets',
     'src/python/pants/backend/python/tasks',
     'src/python/pants/base:exceptions',
+    'src/python/pants/python',
     'src/python/pants/util:dirutil',
     'src/python/pants/testutil:task_test_base',
   ],
