+++ tests/python/pants_test/backend/python/tasks/test_gather_sources.py
@@ -6,11 +6,11 @@ import os
 from pex.interpreter import PythonInterpreter
 
 from pants.backend.python.interpreter_cache import PythonInterpreterCache
-from pants.backend.python.subsystems.python_setup import PythonSetup
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.backend.python.tasks.gather_sources import GatherSources
 from pants.build_graph.files import Files
 from pants.build_graph.resources import Resources
+from pants.python.python_setup import PythonSetup
 from pants.source.source_root import SourceRootConfig
 from pants.testutil.task_test_base import TaskTestBase
 from pants.util.contextutil import temporary_dir
