+++ tests/python/pants_test/backend/python/tasks/test_resolve_requirements.py
@@ -9,10 +9,10 @@ from pex.interpreter import PythonInterpreter
 
 from pants.backend.python.interpreter_cache import PythonInterpreterCache
 from pants.backend.python.python_requirement import PythonRequirement
-from pants.backend.python.subsystems.python_setup import PythonSetup
 from pants.backend.python.targets.python_requirement_library import PythonRequirementLibrary
 from pants.backend.python.tasks.resolve_requirements import ResolveRequirements
 from pants.base.build_environment import get_buildroot
+from pants.python.python_setup import PythonSetup
 from pants.testutil.task_test_base import TaskTestBase
 from pants.util.contextutil import temporary_dir, temporary_file
